pyproject.toml
[build-system]
requires = ["setuptools>=68.0", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "musktracker"
version = "0.1.0"
description = "Advanced statistical modeling of Elon Musk tweet volume"
requires-python = ">=3.11"
dependencies = [
    "sqlalchemy>=2.0.0",
    "psycopg2-binary>=2.9.0",
    "alembic>=1.13.0",
    "pandas>=2.0.0",
    "numpy>=1.24.0",
    "statsmodels>=0.14.0",
    "scipy>=1.11.0",
    "tweepy>=4.14.0",
    "tenacity>=8.2.0",
    "python-dotenv>=1.0.0",
    "python-dateutil>=2.8.0",
    "pytz>=2023.3",
    "click>=8.1.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.4.0",
    "pytest-cov>=4.1.0",
    "black>=23.0.0",
    "mypy>=1.5.0",
    "ruff>=0.1.0",
]

[project.scripts]
musktracker = "musktracker.__main__:main"

[tool.setuptools.packages.find]
where = ["."]
include = ["musktracker*"]

[tool.black]
line-length = 100
target-version = ["py311"]

[tool.mypy]
python_version = "3.11"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = false

[tool.ruff]
line-length = 100
target-version = "py311"

